miriad: allow bigger UV variable arrays
MAXVAR is the maximum number of UV variables allowed, not the maximum
size of any variable in bytes.

We are still dodgy in `UVObject_rdvr` because we don't check that our buffer
(`value`) can actually fit the variable that is being read, which consists of
`length` items. This is a bit of a hassle to check since to check for overflow
we need to switch over the variable type to know how many bytes each item is.