Rename `aipy_src` to `aipy`.
Clearly there was some kind of conflict that motivated the original gymnastics
with naming this directory, but having the directory name match the package
name is how *everyone* does things now, and I'm running into problems with pip
development installs due to the name mismatch. Let's change the name to the
standard form and see what happens.